requirements.txt
huggingface_hub 
pandas 
matplotlib 
seaborn
numpy
torch
transformers
datasets
pandas
pyyaml
jupyter